LockFreeQueue
Stress Test로 성능 측정 예정

+++ SERVER/VIEW/SERVER/SERVER/multi_iocp_server.cpp
@@ -445,7 +445,7 @@ void process_packet(const int c_id, char* packet)
 	case CS_INTERACTION: {
 		CS_INTERACTION_PACKET* p = reinterpret_cast<CS_INTERACTION_PACKET*>(packet);
 		int cur_stage = CL.cur_stage.load();
-		for (size_t i = 0; i < Key_Items[cur_stage].size(); i++) {
+		for (int i = 0; i < Key_Items[cur_stage].size(); i++) {
 			if (Key_Items[cur_stage][i].m_xmOOBB.Intersects(CL.m_xmOOBB))
 				for (auto& cl : Room) {
 					cl.clear_percentage += Key_Items[cur_stage][i].percent;
